requirements.txt
dash==2.0.0    
dash-extensions==0.0.60    
Flask==2.0.2 
ftfy==6.0.3
image==1.5.33 
imutils==0.5.4  
keras==2.6.0
lxml==4.6.3  
nltk==3.6.5    
numpy==1.19.5
opencv-python==4.5.4.58 
matplotlib==3.4.3    
pandas==1.1.5
Pillow==8.4.0    
plotly==5.3.1    
pytesseract==0.3.8  
pytest-shutil==1.7.0
regex==2021.8.3
scikit-learn==1.0.1  
sentencepiece==0.1.96   
six==1.15.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0 
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.1
transformers==4.8.2